v3.7.4 b707
* changed item-drop and xp-drop multipliers to use double values instead of integer

+++ src/main/java/me/lokka30/levelledmobs/customdrops/CustomDropProcessingInfo.java
@@ -34,7 +34,7 @@ class CustomDropProcessingInfo {
     }
 
     public LivingEntityWrapper lmEntity;
-    int addition;
+    double addition;
     Player mobKiller;
     @NotNull final Map<String, Integer> playerLevelVariableCache;
     boolean isSpawner;
